fast/excel/requirements.txt
pandas>=1.3.0
openpyxl>=3.0.0
opencv-python>=4.5.0
Pillow>=8.0.0
requests>=2.25.0
tqdm>=4.60.0
numpy>=1.19.0
